move searching and sorting panel to separate component
