+++ client/src/pages/Dashboard/Dashboard.jsx
@@ -1,4 +1,10 @@
-import { CoffeeCard, AddCoffeeDialog, ViewCoffeeDialog, EditCoffeeDialog } from './';
+import {
+  SearchSortPanel,
+  CoffeeCard,
+  AddCoffeeDialog,
+  ViewCoffeeDialog,
+  EditCoffeeDialog,
+} from './';
 import { Navbar } from 'components';
 import { isLoggedIn, getAllCoffeeData, getCoffeeData, sortCoffees } from 'helpers';
 import { Grid, Fab, createTheme, ThemeProvider, TextField, MenuItem } from '@mui/material';
@@ -40,7 +46,6 @@ function Dashboard() {
   const [viewDialog, setViewDialog] = useState(false);
 
   let currentSort = useRef('creationDate');
-  let searchBy = useRef('all');
   let duplicateDialog = useRef(false);
 
   const toggleViewDialog = (event, reason) => {
@@ -63,56 +68,9 @@ function Dashboard() {
     setAddDialog(!addDialog);
   };
 
-  const searchOptions = [
-    { value: 'all', label: 'All' },
-    { value: 'name', label: ' Log Name' },
-    { value: 'coffeeName', label: 'Coffee Name' },
-    { value: 'coffeeRoaster', label: 'Roaster Name' },
-  ];
-
-  const handleSetSearchBy = (value) => {
-    searchBy.current = value;
-    handleSearch(getSearchValue());
-  };
-
   const clearSearch = useCallback(() => {
     document.getElementById('search').value = '';
   }, []);
-
-  const getSearchValue = () => {
-    return document.getElementById('search').value;
-  };
-
-  const handleSearch = (value) => {
-    let searchResults = [];
-    if (searchBy.current === 'all') {
-      searchResults = allCoffeeData.filter(
-        (coffee) =>
-          coffee.name.toLowerCase().includes(value.toLowerCase()) ||
-          coffee.coffeeName.toLowerCase().includes(value.toLowerCase()) ||
-          coffee.coffeeRoaster.toLowerCase().includes(value.toLowerCase())
-      );
-    } else {
-      searchResults = allCoffeeData.filter((coffee) =>
-        coffee[searchBy.current].toLowerCase().includes(value.toLowerCase())
-      );
-    }
-    setVisibleCoffeeData(sortCoffees(searchResults, currentSort.current));
-  };
-
-  const sortOptions = [
-    { value: 'creationDate', label: 'Log Date' },
-    { value: 'name', label: 'Log Name' },
-    { value: 'coffeeName', label: 'Coffee Name' },
-    { value: 'coffeeRoaster', label: 'Roaster' },
-    { value: 'coffeeWeight', label: 'Weight' },
-    { value: 'restDays', label: 'Rest Days' },
-  ];
-
-  const handleSort = (key) => {
-    currentSort.current = key;
-    setVisibleCoffeeData(sortCoffees(visibleCoffeeData, key));
-  };
 
   const updateAllCoffeeData = useCallback(() => {
     getAllCoffeeData().then((coffees) => {
@@ -161,54 +119,13 @@ function Dashboard() {
           <h2 className="text-4xl text-rr-brown-primary font-bold text-center py-8">
             Coffee Dashboard
           </h2>
-          <Grid container spacing={1} className="mb-4">
-            <Grid item xs={3} sm={2}>
-              <TextField
-                select
-                fullWidth
-                id="searchBy"
-                label="Search"
-                variant="outlined"
-                color="primary"
-                defaultValue="all"
-                onChange={(e) => handleSetSearchBy(e.target.value)}
-              >
-                {searchOptions.map((option) => (
-                  <MenuItem key={option.value} value={option.value}>
-                    {option.label}
-                  </MenuItem>
-                ))}
-              </TextField>
-            </Grid>
-            <Grid item xs={9} sm={4}>
-              <TextField
-                fullWidth
-                id="search"
-                label="Search"
-                variant="outlined"
-                color="primary"
-                onChange={(e) => handleSearch(e.target.value)}
-              />
-            </Grid>
-            <Grid item xs={12} sm={6} className="inline-flex justify-end">
-              <TextField
-                select
-                className="w-full sm:w-40"
-                id="sortBy"
-                label="Sort"
-                variant="outlined"
-                color="primary"
-                defaultValue="creationDate"
-                onChange={(e) => handleSort(e.target.value)}
-              >
-                {sortOptions.map((option) => (
-                  <MenuItem key={option.value} value={option.value}>
-                    {option.label}
-                  </MenuItem>
-                ))}
-              </TextField>
-            </Grid>
-          </Grid>
+          <SearchSortPanel
+            allCoffeeData={allCoffeeData}
+            visibleCoffeeData={visibleCoffeeData}
+            setVisibleCoffeeData={setVisibleCoffeeData}
+            currentSort={currentSort}
+            sortCoffees={sortCoffees}
+          />
           {visibleCoffeeData.some((coffee) => coffee.isPinned) && (
             <>
               <h2 className="text-sm text-rr-brown-primary font-bold uppercase ml-2 mb-1">
