+++ client/src/pages/Dashboard/index.js
@@ -1,3 +1,4 @@
+export { default as SearchSortPanel } from './SearchSortPanel';
 export { default as CoffeeCard } from './CoffeeCard';
 export { default as CoffeeCardMenu } from './CoffeeCardMenu';
 export { default as AddCoffeeDialog } from './Dialogs/AddCoffeeDialog';
